First step towards making DHMQC into a proper python application.
This commit restructures the DHMQC package completely and introduces
a setup.py script which can be used to install the package. At the
moment it doesn't compile the C-code, it only adds a CLI entry point
called 'dhmqc'. For now the dhmqc CLI is limited to creating tile
coverages, or tindexes, as they are now called.

Everything else is broken. Will be fixed in coming commits.